Add VcsAnnotationItemDelegate, for control of rendering and tooltip
Summary:
Uses upcoming KTextEditor::AbstractAnnotationItemDelegate

While staying in the boundaries of line-based rendering,
the delegate takes advantage of knowledge about grouped
lines for the same annotation, and does not repeat the main
info, but writes a secondary info on a second line (taking into
account visibility) and keeps further ones blank.

It also applies an ActiveToolTip for the tooltip on the annotation,
to allow copying data and in the future clicking links in the shown
commit message.

Reviewers: kfunk

Subscribers: croick, kfunk, brauch, apol, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D8709