Don't ask for compiler-specific includes in QmlJS
Summary:
We can't provide these, since there is no QmlJS compiler.

The entire design is a bit problematic, because the QmlJS plugin asks
the DefinesAndIncludesManager for include directories, but that manager
is somehow tied to the idea of using the configured compiler to provide
additional includes and defines. However, the configured compiler is
obviously only used for C/C++ and similar languages, not QmlJS.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D8654