gcclikecompiler: Fix execution of compiler qprocess under windows
Summary:
Rather than passing null device as input file to the compiler,
"-" is passed as argument which makes the compiler take input from
stdin, then qprocess' standard input file is set to QProcess:nullDevice()
to make it read eof.

Note that it fixes standard includes not being passed to language plugin if
clang is set as default compiler under windows.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, brauch, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D8831