Don't add the same targets multiple times for nested CMake projects
When we import a cmake project folder and that contains nested
CMake projects, we ended up showing the same target multiple times,
since the CMake server output will reference the same targets for
the parent project and then again for the nested project(s).

Ensure the targets are only added once to fix this issue.

BUG: 387095