compute itemrepository version from version plus KDEVELOP_VERSION
This ensures that for each release, we clear the item repository.
In the past, we had various issues related to old caches -- this change
should avold that, since we no longer need to remember to bump the
ITEMREPOSITORY_VERSION variable on each release.

Changing ITEMREPOSITORY_VERSION is still possible to force a clear even
if KDEVELOP_VERSION is untouched.