Use custom enum MLLV::LastLevelViewMode, not raw KSPM::FilterBehavior
Summary:
While FilterBehavior is used for the implementation, it's full range
of enum values does not make sense in the API of MultiLevelListView.
Instead some custom enum LastLevelViewMode describing the actual behaviour
is introduced, and the setter method renamed accordingly.
And KSelectionProxyModel is properly hidden away as implementation detail.

Test Plan: Builds and works as before.

Reviewers: #kdevelop, kfunk, brauch

Reviewed By: #kdevelop, kfunk

Subscribers: aaronpuchert, brauch, kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D9316