Format comments before setting them on the DUChain
Summary:
This removes the /// and /* ... */ prefixes. Not only does this
make the stuff way more readable, it also fixes our JSON tests...

Return proper non-C++ specific includes/defines for non-project files

When we query the includes/defines for a file outside of any project
we always returned the C++ compiler includes/defines. This was because
we never tried to guess the language type then, since we operated only
on ProjectBaseItems pointers which are obviously nullptr for files
outside of any project. Instead, operate on QString path
representations like elsewhere in the IADM already. This allows us to
guess the language type even for C, ObjC and Cuda files when they are
not part of a project.

No unit test is added, but I found this while working on a separate
feature that has a proper unit test attached.

Handle _Complex types

We just create a delayed type that holds the string representation.
This is enough to get the type displayed in the tooltips and it also
removes the pesky warnings from the command line about unhandled
types.

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D9472