Resume background parser when replacing TestCore's project controller
Summary:
This fixes the clang duchain unit tests that set a custom project
controller. Without this patch, the background parser would be
suspended and none of the unit tests would produce a DUChain in time.

I wonder why I haven't spotted this before when introducing the
background parser suspension on startup... Was I not running the
tests? Also note that tons of these tests are actually broken, seems
like I'm not the only one who's not running the tests :(

Reviewers: #kdevelop

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D9722