Performance: Reuse the global icon loader
Summary:
I just found a noticeable GUI hang when showing a problem widget
for a missing include path. There, the profiler pointed me at
KIconLoader which was reconstructed whenever we query for the icon
representing the severity of a problem.

By reusing the global icon loader, we can get the information from
the cache which is significantly faster.

Reviewers: #kdevelop

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D9783