GDB Qt pretty printers: print size of container
Summary:
The pretty printers already had a pretty approach for empty
containers, these got an 'empty ' prepended to their output.
But when you are printing large containers, how do you know the
size of the container? Often, it's not straight forward to get
to the size, e.g. for a QList. To help the user of the debugger,
always append '(size = N)' to the output, which works for all
situations. Remove the empty prefix, which is now obsolete.

Reviewers: #kdevelop, dfaure

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D9918