Properly save state of KTextEditor Views
Summary:
Call the KTextEditor::View methods for reading & storing the state:
- KTextEditor::View::readSessionConfig()
- KTextEditor::View::writeSessionConfig()

Remove our implementation of storing the state of ...:
- Cursor
- Selection

By using the KTextEditor methods we do save the states of ...:
- Cursor
- Dynamic Word Wrap
- Text Folding nodes
- ...

Thus, with this patch the storing & retrieving of the selection state is
lost. To be honest, I don't see that as a super useful feature anyway.

FIXED-IN: 5.3.0
BUG: 385919

Reviewers: mwolff

Reviewed By: mwolff

Subscribers: mwolff, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11601