Only do fuzzy path matching on the very last path/filter segments
Summary:
The old code joined all filter segments and then used that to match
it against the last path segment. This is quite odd behavior, imo:
If I type "foo/bar", then I wouldn't expect it to match a file called
"fooXbar". On the other hand, I would expect "foo/bar" to fuzzy-match
"foo/xBxAxRx", since the "foo" segment is matched and then "bar" is
fuzzy-matched against the "xBxAxRx" segment.

The new tests fuzzy2, fuzzy3 and fuzzy5 fail with the old code but
pass with the new, simplified code.

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11933