Refactor matchPathFilter to enable more advanced result sorting
Summary:
We now return a plain int quality which is then used to sort the
matched filter items. This patch does not touch the sorting
behavior itself - this will be done in follow-up patches.
Most notably, I plan to allow finer grained sorting of the
OhterMatch category, such that matches closer to the end of
the path are preferred.

The benchmark shows no big difference before/after this patch,
giving me confidence that I'm not introducing any big
performance changes.

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11936