Simplify code and remove duplicate loop to find exact matches
Summary:
Sadly, the code is not any faster (nor slower) - according to
the benchmark the performance stays pretty much the same. Still,
it's worth it to merge the loop for readability I think. It clearly
shows that we want to go through the segments once to find a
certain match quality.

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11938