Penalize matches in the shared project prefix path
Summary:
This is a refinement of the previous patch, which isn't perfect yet:
Assume you have a project at /A/B/C. Now, you have
these two files in there:

/A/B/C/D/file
/A/B/C/C/E/file

Note also how quickopen trims the shared prefix path when it displays
the results:

D/file
C/E/file

When you search for C/file, both files used to get the same quality,
even though the former one matches only in the shared prefix path
of the project.

This patch adds a penalty when the match happens in the shared
prefix. Thus, when you search for C/file you'd now get the second
path preferred. This is imo much more desirable.

The benchmark timings are still pretty much the same.

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D11940