Use the project name more consistently
The user-provided project name will now also be used to name the .kdev4
project files. This makes it possible to create multiple KDevelop
projects in a single directory from within KDevelop.
The project import process is changed slightly: the dialog questioning
the user if an existing project file should be used or overridden will
now only appear when importing a non-kdev4 "project" file to a project
name that is already used in the same directory.

Differential Revision: https://phabricator.kde.org/D9344