Allow overriding the path to the builtin clang compiler headers
The current code to detect the path for the builtin clang compiler
headers fails on Gentoo. Allow setting the KDEV_CLANG_BUILTIN_DIR
environment variable for such cases as a stop-gap measure to
workaround this for now.

We will have to find a better approach to this problem, it seems.
Most notably, we will probably have to detect the path at runtime
based on the libclang we actually use then. This may be different
from the one we compiled against...

For now, we also report an error (only visible on the command line)
when the KDEV_CLANG_BUILTIN_DIR path is "wrong", i.e. doesn't contain
the varargs.h header. In such cases, we disable the kdev-clang plugin,
since it would be unusable anyways.

CCBUG: 393779