Use "edit-delete" icon for buttons which delete data
Summary:
The "list-remove" icon does not signal the user that the related action
actually will results in non-recoverable data deletion. Removing an item
from a list does not necessarily mean the item is destroyed.

Using "edit-delete" instead makes the deletion effect more obvious to the
user. In the Breeze icon theme this also means the related signal color is
used to point out the nature of the action.

Reviewers: #kdevelop, aaronpuchert

Reviewed By: #kdevelop, aaronpuchert

Subscribers: aaronpuchert, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14304