Fix GdbTest::testUpdateBreakpoint from messed-up break points positions
Summary:
Commit e4b139268ccf9be93b0049c503d4caab6241b4ec added some lines to the
sources of the debugged code, but only updated the first breakpoint
position, with the second manually inserted suddenly being before the
first.
This results in a race condition, given the command for the
second breakpoint is only executed when the actual program is also run
and might have already reached the originally set breakpoint. In that case
the manually set breakpoint no longer will be reached, breaking the
assumptions of the test case.

Remove also extra session->stepInto(); which served no obvious purpose
for the given test case.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, arrowd, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14337