Use platform-specific default shortcuts in the  document switcher
This introduces a platform-specific shortcutAccelerator variable
that is set to Qt::CTRL except when running on Mac and the
DontSwapCtrlAndMeta attribute is set. In the latter case the variable
is set to Qt::ALT (instead of Qt::META).
This prevents triggering the system application switcher instead of
the document switcher on Mac (and presumably MS Windows).

Differential Revision: https://phabricator.kde.org/D6060