cppcheck: pass "-i" & path as separate argument items, not as single string
Summary:
Passing "-i /some/path/CMakeFiles" as single string with the args for the
cppcheck invocation results in the string being passed quoted due to the
space inside, '-i /some/path/CMakeFiles'. Seems this still results in the
wanted processing, but it looks strange in the console log.

Test Plan: cppcheck still works as before and skips the CMakeFiles subfolder.

Reviewers: #kdevelop, antonanikin, apol

Reviewed By: #kdevelop, apol

Subscribers: kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14675