Store sizeof and friends as numbers instead of comment.
Summary:
This way they can be used by other plugins, for example to visualize padding
in structures. They also take a bit less space in duchain in this form.
It is not longer necessary to exclude them from the json test run, since
they are not part of the comment.

The end result is now the same - the values are shown as text in tooltip.

Test Plan:
Verify that message similar to "size: 96 Bytes; aligned to: 8 Bytes" is
shown in tooltip for classes and class members.

Reviewers: #kdevelop, aaronpuchert

Reviewed By: #kdevelop, aaronpuchert

Subscribers: brauch, aaronpuchert, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14479