Generate strings with lists by collecting items, then use QSL::join()
Saves memory reallocations and copying on every appending to the string
collecting the result, once for the item and once for some possible
separator.

Ideally we would just reserve enough space in the target string and
append the new content in there directly. But that requires knowning the
final size in advance or the maximal possible, so the