Fix VariableToolTip auto resizing.
Summary:
The VariableToolTip is supposed to resize vertically as much as can fit
in the window whenever the content changes. It was not working because
the m_itemHeight was 0 which was because the QTreeView::indexRowSizeHint
was called with index from m_model instead of m_proxy.

Same for the initial selection below.

Test Plan:
Debug some program, stop execution and hover over some variable. Expand
some subitems - the tooltip should be resizing.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14879