Fix kdev_perforce unittest to run on kde windows CI and everywhere else.
Summary:
Fix unittest properly. The previous 'fix' for the unittest did not work on CI.

And IMHO it was not too pretty. This one should work on the CI and everywhere else for that matter.

Test Plan: Tested on Linux and windows builds. If this works unittest will pass on KDE Windows CI as well.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14958