make AStyleStringIterator::getPeekStart() return valid values
The astyle v3 API introduced a new virtual ASSourceIterator::getPeekStart()
method that has to be implemented in classes inheriting ASourceIterator.
Currently that method is only used in an assert that verifies whether it
returns a valid (i.e. positive) value. This was overlooked in the upgrade
of the bundled astyle library code.

BUG: 399048
Differential Revision: https://phabricator.kde.org/D15605