Add missing break in QmlJs code completion
The values operator{Start,End} are used in functionCallTips to find the
m_typeToMatch. The idea is that if we write a + then we want something
of the same type as a to add to it.

With the fallthrough we apply this for commas as well, so we try to
match the type before the comma. But function arguments generally do not
have the same type.

This change was separated from the changes in D15694.