Make implicit fallthroughs a compiler error, remove unneeded breaks
Summary:
With -Wimplicit-fallthrough the compiler warns about unannotated
fallthroughs in switch statements. Since it is supported by both GCC and
Clang, and can easily be fixed by adding Q_FALLTHROUGH() we enable it as
error.

As a consequence, we don't need to add redundant break statements as
safety measure. So we also warn about that, but not as error, especially
since it's only supported on Clang. (As far as I know.)

Reviewers: #kdevelop, kossebau, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: mwolff, apol, kfunk, rjvbb, brauch, mssola, kossebau, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D15694