Make test outputmodel pass on windows.
Summary: When doing the benchmark test on windows. Make sure that we use windows paths. Otherwise the test will fail (in debug builds) in an assert in util/path.cpp

Test Plan: Run the test in a Windows debug build environment. Before patch it fails. After patch is applied it succeeds. Run test on Linux and see that test stil completes

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D17077