Make the filteringstrategy test work on Windows as well.
Summary: This requires some changes to some of the regexes used in CompilerFilterStrategy to get to work. Also the test itself is now somewhat more specific about what path type to test on which platform.

Test Plan: Compile the changes on Windows and verify that the test now passes (In Debug mode). Compile the changes on Linux and verify that all tests pass (as before). Verify that the bench number still look reasonable

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D17098